pyproject.toml
[project]
name = "image-processor-for-shrimp"
version = "0.1.0"
description = "Image processor for Shrimp"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pillow>=8.0.0",
]

[dependency-groups]
dev = [
    "image-processor-for-shrimp",
    "pyinstaller>=6.16.0",
]

[tool.uv.workspace]
members = [
    ".",
]

[tool.uv.sources]
image-processor-for-shrimp = { workspace = true, editable = true }
